requirements.txt
certifi==2021.10.8
pymongo==4.0.2
python-dotenv==0.19.2
selenium==3.141.0